requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
flask-cors==4.0.1
flasgger==0.9.7.1
python-dotenv==1.0.1
apscheduler==3.10.4
psycopg2-binary==2.9.9
marshmallow==3.21.3